fixing bugs in migration down functions

+++ database/migrations/2022_09_20_132603_change_staking_standard_odd_column_type_to_decimal.php
@@ -26,7 +26,7 @@ return new class extends Migration
     public function down()
     {
         Schema::table('stakings', function (Blueprint $table) {
-            $table->tinyInteger('standard_odd')->change();
+            $table->Integer('standard_odd')->change();
         });
     }
 };
